Remove project enterprises completely, implement edit and view enterprises
