+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/constants.ts
