--- frontend/src/components/manage/Blocks/ProjectsListing/ProjectsEnterprises/create/PEnterpriseHeader.tsx
@@ -1,84 +1,0 @@
-import { useState } from 'react'
-
-import HeaderTitle from '@ors/components/theme/Header/HeaderTitle'
-import { PageHeading } from '@ors/components/ui/Heading/Heading'
-import EnterpriseCancelButton from '../../Enterprises/create/EnterpriseCancelButton'
-import PEnterpriseEditActionButtons from '../edit/PEnterpriseEditActionButtons'
-import PEnterpriseCreateActionButtons from './PEnterpriseCreateActionButtons'
-import { PageTitle } from '../../HelperComponents'
-import { EnterpriseStatus } from '../FormHelperComponents'
-import {
-  EnterpriseData,
-  EnterpriseHeaderProps,
-  PEnterpriseType,
-} from '../../interfaces'
-
-import { CircularProgress } from '@mui/material'
-import cx from 'classnames'
-
-const PEnterpriseHeader = ({
-  mode,
-  enterprise,
-  ...rest
-}: EnterpriseHeaderProps & {
-  mode: string
-  enterpriseData: EnterpriseData
-  enterprise?: PEnterpriseType
-}) => {
-  const [enterpriseName, setEnterpriseName] = useState(
-    enterprise?.enterprise?.name ?? '',
-  )
-  const [isLoading, setIsLoading] = useState<boolean>(false)
-
-  const isEdit = mode === 'edit' && !!enterprise
-
-  return (
-    <HeaderTitle>
-      <div className="align-center flex flex-wrap justify-between gap-x-4 gap-y-4">
-        <div className="flex flex-col">
-          <EnterpriseCancelButton
-            type="project enterprise"
-            mode="redirect"
-            isEdit={isEdit}
-          />
-          <PageHeading>
-            {isEdit ? (
-              <PageTitle
-                pageTitle="Edit project enterprise"
-                projectTitle={enterpriseName}
-              />
-            ) : (
-              'New project enterprise submission'
-            )}
-          </PageHeading>
-          {isEdit && <EnterpriseStatus status={enterprise.status} />}
-        </div>
-        <div
-          className={cx('ml-auto flex items-center gap-2.5', {
-            'mt-auto': mode === 'add',
-          })}
-        >
-          <div className="flex flex-wrap items-center justify-end gap-2.5">
-            <EnterpriseCancelButton
-              type="project enterprise"
-              mode="cancel"
-              isEdit={isEdit}
-            />
-            {mode === 'add' ? (
-              <PEnterpriseCreateActionButtons {...{ setIsLoading, ...rest }} />
-            ) : (
-              <PEnterpriseEditActionButtons
-                {...{ enterprise, setIsLoading, setEnterpriseName, ...rest }}
-              />
-            )}
-          </div>
-          {isLoading && (
-            <CircularProgress color="inherit" size="30px" className="ml-1.5" />
-          )}
-        </div>
-      </div>
-    </HeaderTitle>
-  )
-}
-
-export default PEnterpriseHeader
